requirements.txt
pyinstaller
python-github